+++ src/senaite/core/upgrade/v02_07_000.py
@@ -39,6 +39,7 @@ from senaite.core.schema.addressfield import PHYSICAL_ADDRESS
 from senaite.core.schema.addressfield import POSTAL_ADDRESS
 from senaite.core.setuphandlers import add_catalog_column
 from senaite.core.setuphandlers import add_catalog_index
+from senaite.core.setuphandlers import add_dexterity_items
 from senaite.core.upgrade import upgradestep
 from senaite.core.upgrade.utils import UpgradeUtils
 from zope.component import getMultiAdapter
@@ -208,6 +209,7 @@ def remove_at_portal_types(tool):
     logger.info("Remove AT types from portal_types tool ... [DONE]")
 
 
+@upgradestep(product, version)
 def migrate_contacts_to_dx(tool):
     """Migrate Contact objects from Archetypes to Dexterity
     """
@@ -486,3 +488,33 @@ def to_dx_address(value, address_type=NAIVE_ADDRESS):
         "subdivision2": u(value.get("district") or ""),
         "country": u(value.get("country") or ""),
     }
+
+
+@upgradestep(product, version)
+def create_setup_contacts_folder(tool):
+    """Create the Contacts container in the setup folder
+    """
+    logger.info("Creating Contacts container in setup folder ...")
+
+    # run required import steps
+    tool.runImportStepFromProfile(profile, "typeinfo")
+    tool.runImportStepFromProfile(profile, "actions")
+
+    setup = api.get_senaite_setup()
+
+    # Check if contacts folder already exists
+    if not setup.get("contacts"):
+        items = [("contacts", "Contacts", "Contacts")]
+        add_dexterity_items(setup, items)
+        logger.info("Contacts container created")
+    else:
+        logger.info("Contacts folder already exists [SKIP]")
+
+    logger.info("Creating Contacts container in setup folder [DONE]")
+
+
+@upgradestep(product, version)
+def notify_upgrade(context):
+    """Dummy func to force the call of before and after upgrade events
+    """
+    pass